First working re-factoring version with small issues.
